+++ src/Pages/Catalog/Catalog.module.scss
@@ -10,81 +10,10 @@
     gap: 2.5rem;
     padding: 2.5rem 0;
 
-    .title {
-      position: relative;
-      display: flex;
-      justify-content: space-between;
-
-      h2 {
-        font-family: fonts.$third-family;
-        font-size: 2rem;
-        font-weight: 700;
-        line-height: 0.9;
-        color: colors.$grey-300;
-        letter-spacing: 0.08em;
-      }
-
-      .sort {
-        position: absolute;
-        z-index: 5;
-        top: 0;
-        right: 0;
-
-        background-color: colors.$accent-white;
-        border-radius: 0.75rem;
-      }
-
-      .size {
-        position: absolute;
-        z-index: 5;
-        top: 0;
-        right: 500px;
-
-        background-color: colors.$accent-white;
-        border-radius: 0.75rem;
-      }
-    }
-
     .main {
       display: grid;
       grid-template-columns: 282px 1fr;
       gap: 5.25rem;
-
-      .filters {
-        height: fit-content;
-        padding: 1.5rem 1rem;
-        background-color: colors.$accent-white;
-        border-radius: 0.75rem;
-      }
-
-      .box {
-        position: relative;
-        display: flex;
-        flex-direction: column;
-        gap: 2rem;
-
-        .books {
-          z-index: 2;
-          display: grid;
-          grid-gap: 1.5rem;
-          grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
-
-          &.disabled {
-            opacity: 0.5;
-          }
-        }
-
-        .fetching {
-          position: absolute;
-          z-index: 10;
-
-          width: 100%;
-          height: 100%;
-
-          opacity: 0.5;
-          background-color: colors.$grey-100;
-        }
-      }
     }
   }
 }
