+++ src/Pages/Catalog/CatalogHeader/CatalogHeader.module.scss
@@ -1,0 +1,37 @@
+@use '@/styles/colors.scss';
+@use '@/styles/fonts.scss';
+
+.title {
+  position: relative;
+  display: flex;
+  justify-content: space-between;
+
+  h2 {
+    font-family: fonts.$third-family;
+    font-size: 2rem;
+    font-weight: 700;
+    line-height: 0.9;
+    color: colors.$grey-300;
+    letter-spacing: 0.08em;
+  }
+
+  .sort {
+    position: absolute;
+    z-index: 5;
+    top: 0;
+    right: 0;
+
+    background-color: colors.$accent-white;
+    border-radius: 0.75rem;
+  }
+
+  .size {
+    position: absolute;
+    z-index: 5;
+    top: 0;
+    right: 500px;
+
+    background-color: colors.$accent-white;
+    border-radius: 0.75rem;
+  }
+}
