+++ src/Pages/Catalog/Products/Products.module.scss
@@ -1,0 +1,31 @@
+@use '@/styles/colors.scss';
+@use '@/styles/fonts.scss';
+
+.box {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+
+  .books {
+    z-index: 2;
+    display: grid;
+    grid-gap: 1.5rem;
+    grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
+
+    &.disabled {
+      opacity: 0.5;
+    }
+  }
+
+  .fetching {
+    position: absolute;
+    z-index: 10;
+
+    width: 100%;
+    height: 100%;
+
+    opacity: 0.5;
+    background-color: colors.$grey-100;
+  }
+}
